always honor newseed text input. also add variety but predictability to the world each day
the floor generation should ALWAYS honor the text field input.
even on page load. otherwise its quite misleading. you think you are running the same track as your friends but you are not unless you clicked "Go!" at least once.

found a way to add variety to the tracks every day but also ensure that people can automatically be running the same track as others.
what good is competition if there is no confidence that we are running that same race/track?
_dL